added authentication

+++ src/main/java/com/vss/wardrober/controllers/PieceController.java
@@ -3,10 +3,12 @@ package com.vss.wardrober.controllers;
 import com.vss.wardrober.DTOs.PieceDTO;
 import com.vss.wardrober.models.PieceModel;
 import com.vss.wardrober.services.PieceService;
+import com.vss.wardrober.services.UserService;
 import jakarta.validation.Valid;
 import org.springframework.beans.BeanUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -18,9 +20,11 @@ import java.util.Optional;
 public class PieceController {
 
     private final PieceService pieceService;
+    private final UserService userService;
 
-    public PieceController(PieceService pieceService) {
+    public PieceController(PieceService pieceService, UserService userService) {
         this.pieceService = pieceService;
+        this.userService = userService;
     }
 
     @GetMapping
@@ -39,36 +43,56 @@ public class PieceController {
 
     @PostMapping
     public ResponseEntity<PieceModel> postPiece(@Valid
-                                                @RequestBody PieceDTO pieceDTO) {
+                                                @RequestBody PieceDTO pieceDTO,
+                                                JwtAuthenticationToken token) {
+
+        var user = userService.findById(Long.parseLong(token.getName()));
+
         var pieceModel = new PieceModel();
         BeanUtils.copyProperties(pieceDTO, pieceModel);
+        pieceModel.setUserModel(user.get());
         return ResponseEntity.status(HttpStatus.CREATED).body(pieceService.save(pieceModel));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deletePiece(@PathVariable Long id) {
+    public ResponseEntity<String> deletePiece(@PathVariable Long id,
+                                              JwtAuthenticationToken token) {
+
+
         Optional<PieceModel> piece = pieceService.findById(id);
         if (piece.isEmpty()){
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "piece not found");
         }
-        pieceService.deleteById(id);
+
+        if (piece.get().getUserModel().getId().toString().equals(token.getName())) {
+            pieceService.deleteById(id);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         return ResponseEntity.status(HttpStatus.OK).body("Piece Deleted");
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<PieceModel> putPiece(@Valid
                                                @PathVariable Long id,
-                                               @RequestBody PieceDTO pieceDTO) {
+                                               @RequestBody PieceDTO pieceDTO,
+                                               JwtAuthenticationToken token) {
 
         Optional<PieceModel> piece = pieceService.findById(id);
         if (piece.isEmpty()){
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "piece not found");
         }
-        var pieceModel = new PieceModel();
-        BeanUtils.copyProperties(pieceDTO, pieceModel);
-        pieceModel.setId(piece.get().getId());
-        pieceModel.setUserModel(piece.get().getUserModel());
-        return ResponseEntity.status(HttpStatus.OK).body(pieceService.save(pieceModel));
+        if (piece.get().getUserModel().getId().toString().equals(token.getName())) {
+            var pieceModel = new PieceModel();
+            BeanUtils.copyProperties(pieceDTO, pieceModel);
+            pieceModel.setId(piece.get().getId());
+            pieceModel.setUserModel(piece.get().getUserModel());
+            return ResponseEntity.status(HttpStatus.OK).body(pieceService.save(pieceModel));
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
     }
 }
